Move CapturingLogger to the test target
`CapturingLogger` is only ever instantiated by tests
(`AccessorOperatorsTests`, `EvaluatorTests`, `LoggerTests`); having it
in the production module was a holdover from an earlier iteration where
production helpers needed to reference it. Move the type to
`Tests/RulesEngineTests/Helpers/CapturingLogger.swift` and drop the
stale "lives in the production module so non-test callers can reference
it" doc comment. The legacy `RevenueCat.xcodeproj` gets the file
registered in the same four pbxproj spots used for the existing
helpers; Tuist regenerates from globs.
