+++ Tests/RulesEngineTests/Helpers/CapturingLogger.swift
@@ -1,0 +1,31 @@
+//
+//  CapturingLogger.swift
+//
+//  Created by Antonio Pallares.
+//
+
+import Foundation
+
+@testable import RulesEngine
+
+/// Test-only logger that records every warning it receives for later
+/// assertion. `final` because there is no reason to subclass it.
+final class CapturingLogger: RulesEngineLogger {
+
+    private let lock = NSLock()
+    private var captured: [String] = []
+
+    init() {}
+
+    var warnings: [String] {
+        lock.lock()
+        defer { lock.unlock() }
+        return captured
+    }
+
+    func warn(_ message: String) {
+        lock.lock()
+        defer { lock.unlock() }
+        captured.append(message)
+    }
+}
